[Shader Graph] Simplified the code for Camera Direction
This is a Shader Graph optimization. It simplifies the HLSL code used to generate the Camera View Direction. The output result is the same, but it now requires no math whereas the previous version was doing several matrix operations.

Also included is an update to the documentation to reflect the change.

Fixes this Jira issue: https://jira.unity3d.com/browse/PLATGRAPH-3720