Move a copy of main manual SRP Batcher and BRG pages to HDRP
During the URP migration, we've needed to move the SRP Batcher and BRG pages to a URP section.
However this means there's no natural place for HDRP users to look for the content.
So (in line with how we've treated URP+HDRP docs elsewhere), this PR creates a copy of the docs in the HDRP manual.

It's a direct copy of the URP version, but with any references to URP removed.

Please **don't** review the actual content. We know it needs improvement and fixes but that's out of scope for the migration. We just want to make sure the pages carry over to the HDRP manual ok.

Jira ticket: https://jira.unity3d.com/browse/DOCG-5948