[UUM-78038] Fix cascade shadows and distance shadowmask not blending
UUM-78038

There were issues where blending would occur further than the maximum distance for cascade shadows and could be broken depending on the camera view.

The alpha value of EvalShadow_GetSplitIndex is intended for blending between cascades. This is also used where the last cascade and shadowmask are blended, which is why the blending is not working properly.

Modified to blend cascade shadows and shadowmask in the same way as URP.

The images below compare the modified HDRP and URP.

* Blending of cascade shadows and shadowmask.
![image](https://media.github.cds.internal.unity3d.com/user/1641/files/fb63bbbc-c8fc-410d-ae0f-8dfffa22bcbd)

* To visualize the blending, cascade shadows are represented as 0 (black area) and shadowmask as 0.5 (gray area). Blending occurs at the maximum distance of cascade shadows.
![image](https://media.github.cds.internal.unity3d.com/user/1641/files/657f732a-934d-4362-b051-5f8f8ce8785b)

Comparison of original HDRP and modified HDRP as the camera distance increases. You can see that the blending works properly regardless of the distance.

* Original
![image](https://media.github.cds.internal.unity3d.com/user/1641/files/11e28622-fcc4-468e-be0d-b74f4014cd99)

* Modified
![image](https://media.github.cds.internal.unity3d.com/user/1641/files/98b11973-7485-452c-975f-e39034f7721e)