[Port] [2022.3] Fix a freeze when the Camera farClipPlane is set to Infinity
This PR fixes a freeze when the `Camera farClip` is set to `Infinity`.

Basically, the Light Forward+ algorithm was freezing the editor because the farClip becomes negative, which is not a valid scenario, so most of the calculations are not ending (freeze).

Clamping it to a minimum value of 0 fixes the freeze without any drawback.