[WebGPU] URP Foundation Yamato Tests
https://jira.unity3d.com/browse/PLATGRAPH-3702

The purpose of this PR is to introduce WebGPU yamato jobs for the URP Foundation test project / suite. With this change, we will be able to run yamato jobs for webgpu URP off of trunk, and eventually get these jobs into QV with the appropriate amount of green test runs. Currently, there are around ~10 tests that are still failing that are gated behind a test filter flag in the metafiles. They are documented and will enabled one by one once the underlying issues are fixed, but this PR is focusing on the yamato jobs themselves rather than fixing all of the graphics tests.

Currently, this does not run the GPU driven tests, only standalone but it does run both render graph and non render graph versions