Lighting/fix freeze during apv bake in terrain demo scene
https://jira.unity3d.com/browse/UUM-73316

Editor is not frozen, the bake is just taking an insane amount of time (~2 weeks) because the APV volume added by default covers a huge area.
As a fix: Added a cancelable progress bar to inform the user of what's going on and give him/her a way to abandon the bake.