[HDRP] Fix an expensive call to FindObjectsOfTypeAll causing inspector to be slow in volume inspector UI
Following this recent PR : https://github.cds.internal.unity3d.com/unity/unity/pull/51061
now a lots of override calls the EnsureFrameSettings, which also calls the GetDisplayedCameras() function. 
This function contained an expensive call to FindObjectsOfTypeAll to retrieve the list of all gameview that wasn't really needed for the cost it added. 

This PR removes this call and only get the first gameview which is enough in our use case given that the scene views are alreadt all retrieved as well.