Graphics/SRP/RPF - [UUM-76816] - Fix RenderGraph Error when resizing the sceneview in URP 2D
Current PR fixes an issue where resizing the SceneView in URP2D would spam Render Graph errors in the console. 

When reducing the SceneView to its minimal size in the Editor, we end up with a target pixel size of 1x1 which is valid. However, in URP 2D, once multiplied by a `lighRenderTextureScale` value inferior to 1.0f, this corner case value ends up as 0x0 once casted from float to integer. Safety layers in Render Graph API detect it and throw an error.

Current proposal is to simply add safeguards in URP 2D preventing the dimensions from dropping below 1x1 during the downscale process, in a similar fashion to what was done for URP 3D in #53231. That same PR also added a warning in cross-renderer render pipeline level that will trigger if there is a request for rendering with a 0x0 pixel size.