Fix depth stencil format for cameraDepthAttachment and cameraDepthTexture in URP
Fixes two issues with the depthStencilFormat in URP RenderGraph.

Issue 1 :  (UUM-79752) The depth stencil format in the URP asset settings is ignored for the cameraDepthAttachment and the cameraDepthTexture. The default format is alway used. Passing the right depthStencilFormat is broken in multiple places.

Issue 2: A TextureHandle still uses depthBits instead of the explicit depthStencilFormat. This limits what can be expressed and also introduces a range of bugs when new TextureHandles are being created using a descriptor from an existing handle. 

After fixing Issue 1, the formats are now correct:
![image](https://media.github.cds.internal.unity3d.com/user/1031/files/50df13d2-d345-4ef6-b65e-036158ab084d)

Many additional bugs are fixed in this PR. And the code has been made to be more robust to avoid bugs in the future.