[VFX] Custom hlsl parsing issues
Jira: UUM-79389

### First issue
When this HLSL function is used, the `position` attribute is not recognized
```hlsl
void Test(inout VFXAttributes attributes, in float value, in float altSize)
{
    if (attributes.position.y >= value)
    {
        attributes.size = altSize;
    }
}
```
In this simple case it does no harm, but in more complex cases it can break the shader compilation.

### Second issue
The same function, but with a different formatting is not properly parsed, leading to function's parameters not being identified.
```hlsl
void Test(inout VFXAttributes attributes
   , in float value
   , in float altSize
)
{
    if (attributes.position.y >= value)
    {
        attributes.size = altSize;
    }
}
```