[VFX] Strip tangent not properly computed when using Shader Graph output
To compute the tangents for strips, we check the distances to previous and next particles, and obtain the direction.
When the particles are too close, the direction may not be representative, so we check that they are under a certain threshold.
The threshold used is good enough for distance but, since we are testing distances squared, it turns out to be to large.
This was already fixed for VFX graph generated shaders, but we missed the same code path for Shader Graph generated shaders.

Wrong
<img width="653" alt="image" src="https://media.github.cds.internal.unity3d.com/user/2805/files/cf2712ec-be64-4838-b362-adb0b380e117">

Right
<img width="584" alt="image" src="https://media.github.cds.internal.unity3d.com/user/2805/files/1324fa06-571e-4e8e-a51a-49cfbf70af39">