Auto initialize lighting toggle.
JIRA: https://jira.unity3d.com/browse/GFXFEAT-381

This PR adds a toggle to the Lighting Window which, when enabled, will ensure that the scene has baked lighting data. If lighting data is missing, it will be baked upon opening the scene. We are primarily adding this due to requests from internal users and external asset developers/maintainers - they want to ship their sample scenes without baked lighting data, and have the bake happen on the users machine. They were using auto generate mode for this before, which no longer exists. 

This mode is much simpler than auto generate mode, in that it only kicks in in 1 scenario:
1. You are opening a scene.
2. The scene has no baked lighting data.
3. The scene has "Auto Initialize Lighting" enabled.

If all the above are true, bake the scene.

The new toggle is used as an upgrade path for auto generate mode. Scenes that were using auto before will have this flag enabled. By default it is disabled for new scenes.

![image](https://media.github.cds.internal.unity3d.com/user/2392/files/bfadec06-f973-45b1-a0cf-6e7370a7f40e)