Fixed potential memory leak when cleaning up `ProbeReferenceVolume`
This PR fixes an issue that there is a case where some of the allocated NativeArrays do not get cleaned up properly when cleaning up a `ProbeRefenceVolume`. This can happen when recompiling scripts with APV running.

This PR also includes a minor fix for removing a GC allocation when sorting index defragmenting.

Jira: UUM-82505