[WebGPU] URP Lighting and Post Processing Tests on Yamato
https://jira.unity3d.com/browse/PLATGRAPH-3794 - URP Post Processing JIRA Task
https://jira.unity3d.com/browse/PLATGRAPH-3793 - URP Lighting JIRA Task

The purpose of this PR is to introduce Yamato jobs for webgpu for the URP Lighting and URP Post Processing test suites. With this change, we will be able to run the jobs off of trunk, and eventually get them into QV once we have 10 passing green tests. Currently, there are about ~3 tests left for post processing and 1 for lighting. These will be gated behind test filters for now, as the goal of this PR is to get the Yamato jobs running. They will be tracked in JIRA and removed from the test filter once fixed. 

Previous PR for the URP Foundation tests here: #53971