Graphics/lighting/uum 79279
Cause:
The WorldLightManager.CollectWorldLights() function collects light information for the scene every frame so that it can be processed for render context.
During this process, it uses the FindObjectsByType<HDAdditionalReflectionData> function to update the reflection probe data.
This is a large part of the load.

Solution:
I decided to statically declare a HashSet on the HDAdditionalReflectionData itself to manage it.
Adding and removing from that HashSet was done by redefining the OnEnable and OnDisable functions to work.
The WorldLightManager just needs to use that HashSet data structure on every frame update.