Allow opening RGV but show a better message.
Remove the disabled menu item that was really confusing for users.

This had to do with the fact that in URP render graph is not available if there is the compatibility mode enabled.

Removing that code and adding a better message. I've not added a new API to know why is not available as is two much work, the user should see the docs and try to enable the Render Graph support by their own.

![imagen](https://media.github.cds.internal.unity3d.com/user/3279/files/4f4f0696-60e5-46c7-9ada-02fde4c1df6e)