GFXFOUND-567 Unifiy RenderGraph Compiler with OpenGLES and OpenGL Core
JIRA: https://jira.unity3d.com/browse/GFXFOUND-567 and https://jira.unity3d.com/browse/GFXFOUND-566

The renderGraph currently uses HDRP's implementation with the OpenGLES and OpenGLCore graphics APIs. This PR makes it possible to use the newer renderGraph compiler

The main error with OpenGL + rendergraph was coming from MSAA missing a resolve texture. The error was that the check for resolve texture was done as if there was no autoresolve so it threw when it didn't find the resolve texture. In OpenGL, autoresolve is handled at the graphics API level so it's normal that there is no resolve texture at the end of the pass. The fix is to skip the check if the platform we are using supports autoresolve