[Graphics Tools][HDRP Wizard] Multiple fixes and UI improvements.
https://jira.unity3d.com/browse/UUM-78839
The thing is that this validation should not show the fix button, hide buttons of validations without a fixer

https://jira.unity3d.com/browse/UUM-77499
Simplify it to a single line.

Improve the UI:

Version shown on the window title:

![imagen](https://media.github.cds.internal.unity3d.com/user/3279/files/7568d8a3-a649-4632-b4b9-d0b188265dc7)

Package manager and show on start are always visible in the toolbar.

![imagen](https://media.github.cds.internal.unity3d.com/user/3279/files/567b330d-145d-4e9f-a27b-d9db96a5e62f)

Helpboxes with button instead of custom helpboxes

![imagen](https://media.github.cds.internal.unity3d.com/user/3279/files/4ee1ed7a-31ef-4406-b395-7e17ff91bb8b)

Splitting sections HDRP, VR and DXR. As other windows an inspectors are doing.
![imagen](https://media.github.cds.internal.unity3d.com/user/3279/files/af543313-dacd-4dd4-a4fb-fdce4e3c6962)

Also, each section has a shortcut to it's documentation:
![imagen](https://media.github.cds.internal.unity3d.com/user/3279/files/1b3dfdfb-394a-4988-980f-75aea71eaa1d)

Having a general settings section.
![imagen](https://media.github.cds.internal.unity3d.com/user/3279/files/66140161-4fcc-4dcb-ab88-48ceb787203b)