Fixed layer index out of range in LayeredLitGUI
Jira: [UUM-71157](https://jira.unity3d.com/browse/UUM-71157)

As reported in JIRA, the 4th layer does not work on LayeredLit material when it is copied asset. This is also reproducible if we set the last layer's material to `None` because when the material is copied, the `GUIDArray` is trimmed. By setting the layersGUID size to 4(`kMaxLayerCount`), we can guarantee that the size of the layers to 4 as it is initialized before.

| Before | After |
| -- | -- |
|![image](https://media.github.cds.internal.unity3d.com/user/3842/files/a42c0064-3754-4f13-a446-0ba8e6712349)|![image](https://media.github.cds.internal.unity3d.com/user/3842/files/d56a263b-f6f0-4da1-84a6-90dd9e3fb1ce)|