[SRP] Fixed Corrupted Sprites UUM-77168
This PR fixes UUM-77168, which was a regression from UUM-62522. The main issue is that the "current render pipeline" can be empty string when no render pipeline is selected (which is a valid case). UUM-62522 added a fix to just check the currently supported subshaders by the current render pipeline, but did not handle the empty string case.

While investigating this fix, there was also a warning caused by including `Core2D.hlsl`. That has been fixed, along with some minor updates to generate better Metal shader code.