Stencil-based cross-fade LOD
This PR adds a new option for cross-fade dithering type, which achieves dithering by stencil test instead of alpha testing. By enabling this option, the shader keyword LOD_FADE_CROSSFADE is stripped in forward/deferred pass and in depth-normal prepass (Limitation: it is not stripped if GPU instancing is enabled or if render graph is disabled). Stencil override is incompatible with this option in the above passes because 2 stencil bits (4 and 8) and the comparison function are overridden to render cross-fading objects.

The cross-fade has 4 transition steps. The following image shows a mesh cross-fading with different transition amounts.
![image](https://jira.unity3d.com/secure/attachment/1467810/image-2024-06-14-15-30-58-888.png)

4 transition steps correspond to 4 stencil values. Fade-in and fade-out complement each other, and it is done by using complementing comparison functions.

There is a warning message if stencil override is used along with this option.

![image](https://media.github.cds.internal.unity3d.com/user/7232/files/38193843-e691-4603-8232-4fde87606e7c)