Fix horizontal deformation with water system deformer
This PR allow horizontal deformation with HDRP water system.
Please note that this option is disabled by default thus won't cause any change unless specifically checked by users. 
This was something that was discussed and tested many months ago during multiple hackweeks and was asked by multiple users (forums and via customer support : https://unity.slack.com/archives/C02EZE47X0T/p1726521015570729)

This allows the SG deformers to output a vector3 instead of a single float for deformation. This is split in two to allow for more flexibility in authoring and scalability: 
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/14e40b0e-47c1-424a-bb43-5485f511459f)

This enables behavior like this:

https://media.github.cds.internal.unity3d.com/user/1764/files/1970fc0c-0082-4923-b7f8-b8151b37bf07

Since the increase in memory cost is not insignificant this is disabled by default in an HDRP setting under Water section
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/ef4ca279-9260-4ddd-9755-98b399a6bc51)

Documentation needs to include some limitation of this when it's enabled, this can be done later since water decals is still not properly documented. (Compability with underwater or CPU simulation)
Jira link for documentation : https://jira.unity3d.com/browse/DOCG-6165