[2d] Add TexelSize as optional Init parameter
This is an additional fix to 77050 that addresses SRP-Batcher incompatibility for Shaders created with ShaderGraph for 2D Renderers.

SRP-Batcher for 2D goes through a Custom Path and does not support TexelSize (_TexelSize) and Offset (_ST). 
However ShaderGraph enforces use of TexelSize whenever Textures are sampled due to initializing TexelSize when constructing UnityTexture2D.
This PR makes it optional just like Offset.