[Port] [6000.0] Enable the use of AllGlobals in the UIPass
This PR fixes a bug found by a client when it was trying to use two partial textures in a custom material at UI Pass, even as they were marked as global, as no other Pass was using it later they was reused by other passes, what throws wrong results in the expected render.

When a render pass can run arbitrary materials from users, then it needs to declare it uses All Globals in the pass, in any other case, the resources could be discarded and reused by other passes.

With the change made in this PR, the UI Pass will enable the UseAllGlobals to avoid this error with custom UI shaders.

Even the use of `AllGlobals` can affect performance, we need this change for correctness.