[Port] [6000.0] Fix for Overlay UI in RG + HDR + FXAA/FSR
Backport of #55062  

This PR aims to fix the issue of the Overlay UI not being displayed on screen, when using RG + HDR Output + FSR/FXAA.
The bug was caused by the missing Shader Keyword HDROverlay in finalPostPass, as the method SetupHDROutput would only set the keyword for the uberPass shader.
When FSR/FXAA are enable on RenderGraph more render passes are created and the rendering UI overlay happens in FinalPost, so now we set the keyword based on the material given.