[VFX] Two Custom hlsl bug fixes
### Bug 1
Jira: UUM-83782
When a custom hlsl block has parameter with the same name as a built-in attribute, the VFX does not compile.

- import the provided [package](https://jira.unity3d.com/secure/attachment/1556961/VFXG_HLSL_AttributeRedefinition.unitypackage)
- open the asset named `VFXG_HLSL_AttributeRedefinition`
- try to compile

Actual results: VFX doesn't compile with errors in the console:

> Shader error in '[VFXG_HLSL_AttributeRedefinition] [Simple Burst] Initialize Particles': redefinition of 'direction' at kernel CSMain at VFXG_HLSL_AttributeRedefinition.vfx(191) (on d3d11)

Expected results: 
User should be able to set input names that match the attributes name.

### Bug 2
Jira: UUM-83676

When creating a Custom HLSL operator with two outputs, and that those two outputs are plugged, then the same function is copied twice in the generated compute shader.

Steps to reproduce:
1. Create a new VFX Graph asset (use the simple loop template)
2. Open this VFX Graph
3. Add a Custom HLSL operator
4. Replace the default code by this one:

>void FloatFunction(in float value, out float a, out float b)
{
  a = value * 2;
  b = value * 3;
}
 
5. Connect the output a and output b to the gravity Force input (X and Y)

Actual results: 
The VFX Graph do not compile because the function is duplicated
Expected results: 
The VFX Graph works fine