[UUM-84967] Make RenderGraphViewer script path lookup robust
Fix the `RenderGraphViewer` logic that resolves script paths when clicking on the `#` buttons to not rely on hard-coded assumptions that will change in UPM 8.8.0.

As described in the Jira issue, this change is necessary for the Package Manager team to avoid introducing a regression when upgrading UPM to 8.8.0 or newer because that version introduces a change in how packages are stored in the Library folder.

Jira: https://jira.unity3d.com/browse/UUM-84967