Graphics: Apply dynamic_branch to Fog related keywords
This PR adds a parameter to the ShaderConfig.cs in the URP Config package to apply dynamic branching to fog keywords (FOG_EXP, FOG_EXP2, FOG_LINEAR). Enabling dynamic branching reduces the shader variant build time by 30%.  

More details about the impact of this change are described in this report.

The parameter `k_DynamicBranchFogKeyword` is zero by default. This means dynamic branching is disabled until users change this parameter because this change could cause uncertain issues.

### Additional note

This PR doesn’t provide a way to test dynamic branching on Unity CI, which means we won’t be able to notice the problem in the future. But is it necessary to run all the graphics tests to verify the behavior of a single parameter of the URP Config? Tradeoffs must be considered to address this issue.