[Port] [6000.0] [VFX] Render last particle in particle outputs with strip data
When rendering particle strips, the last particle is ignored. This is because N particles produce N-1 strip segments.
However, if the particle strip system is connected to a regular particle system output, we have to render all particles.

Before, we were adding 1 to the max particle index before testing if the particle must be discarded.
Now, we only do that if the particle is a particle strip output (HAS_STRIPS). We already do similar adjustments in the same function.

The separate path for Shader Graph outputs was already taking this into account, that is why there was a discrepancy.

In the provided repro case, there are 2 particle outputs connected to the same strip system, one with Shader Graph output (left), the other with a regular output(right). The strips behind are only there as a reference.

Before:
![image](https://media.github.cds.internal.unity3d.com/user/2805/files/307ae729-135e-4a39-b2f4-6b0b27f70c00)

After:
![image](https://media.github.cds.internal.unity3d.com/user/2805/files/5b291f82-fa72-4ded-a829-84e04275a8a1)