[Port] [6000.0] Fix APV compatability on GLES
https://jira.unity3d.com/browse/UUM-82539

Adaptive Probe Volume baking doesn't currently work properly on GLES. In this PR, I fix the root cause: The shader that performs jump flooding during probe placement reads and writes to the same buffer in one kernel. This is a race condition that happens to work out on most platforms (jump flooding is eventually-consistent). On GLES, this doesn't work, but not for the reason you might think - GLES simply doesn't support reading and writing to the same buffer in one kernel invocation. By fixing the race, I also fix the GLES incompatability.

While debugging this, I noticed and fixed 3 other issues that were making it very difficult for me to verify my baking fix. Here is a brief description of these:
1. On GLES we use a 4 component unorm texture to store validity. This is because GLES has limited support for single channel unorm. However, we don't account for this when populating the texture, so the data is all messed up. I fixed this by adding a special case when populating.
2. On all mobile targets, we (seemingly accidentally) use half precision reads on the shader side when reading validity. However, validity stores an 8bit integer index in each unorm channel, and must be precisely decoded to work properly. Half precision reads are not sufficient, we need full precision. I changed the textures in question to be bound as full precision on all platforms.
3. Don't initialize compute shaders and enable random write UAVs for scenario blending, if the target doesn't support these.

With that, both baking and sampling of APV seems to work properly on GLES.

There is 1 commit per sub-bugfix.