Fix division by 0, and sphere tile range calculation for forward plus.
This PR fixes a an error in tiling job calculations for URP Forward+ renderer. Tiling job calculates an intersection between an x-z plane and the light volume to determine the range of tiles that the light affects. In the cases where this range falls completely outside of the screen it was incorrectly clamping it to screen boundaries. A previous fix was added for the 'cone' section of spotlight type lights. However the issue also manifests itself in point lights, reflection probes and the cap / 'sphere' section of spotlights.  

Screenshot of issue for this Jira ticket:
<img width="1167" alt="bug" src="https://media.github.cds.internal.unity3d.com/user/760/files/dcd12a08-b35b-4b7e-84d5-f9f2c4017620">

Fixed in this PR:
![image](https://media.github.cds.internal.unity3d.com/user/760/files/d42f9f47-68dd-47b3-aa12-3aaa737ad019)


Light complexity view where the issue can be seen in the 1st and last tile columns (different camera view):
<img width="1166" alt="bugcomplexity" src="https://media.github.cds.internal.unity3d.com/user/760/files/0ed42b26-70c9-409c-bd78-c674f275e143">

Correct light complexity after this PR:
<img width="1165" alt="bugfixcomplexity" src="https://media.github.cds.internal.unity3d.com/user/760/files/c1bafbdd-c1bd-4afc-ab29-1f1961df803d">