[HDRP] Add rolling wave sample scene to water sample
Following this PR : https://github.cds.internal.unity3d.com/unity/unity/pull/52584


https://media.github.cds.internal.unity3d.com/user/1764/files/8365e966-618c-41d8-a7f2-ffd7babc730b



We wanted to give at least one example on how to setup properly a rolling wave given that it's the main reason of this fix. 
This is exactly what this PR does, basically showing one way starting from a modeled mesh in blender and baking that mesh deformation into a texture to be able to read it in the shader graph.