UUM-85566: Backout PR#54311
See [UUM-85566](https://jira.unity3d.com/browse/UUM-85566)
Bisected the issue to this [PR](https://github.cds.internal.unity3d.com/unity/unity/pull/54311). The PR comprises another [bug](https://jira.unity3d.com/browse/UUM-82299) fix, but it is  an optimization and has lower UP. So for now the changes in the PR are reverted and then it can be reimplemented with the repro project in mind.