[VFX] Compilation Optimisation - Speed-up string building
JIRA : https://jira.unity3d.com/browse/VFXG-441

This PR improves the VFX asset compilation time, by changing the way we generate the shaders. Instead of using our old ReplaceMultiline that was slow and allocation-heavy, we recursively collect all the snippets required to build the shader and then perform a series of StringBuilder.Append.
I also tried to lower the amount of Regexes we use, as they are often significantly more expensive than parsing the patterns manually.

Changing the string building algorithm forced me to modify the templates slightly because the order of declaration and usage of some VFX-macros were not compatible.

Performance comparison on 14_PositionBlock.vfx : 
Before : 
![image](https://media.github.cds.internal.unity3d.com/user/2768/files/ba0495db-7501-4ce9-8093-5662355e9bae)
![image](https://media.github.cds.internal.unity3d.com/user/2768/files/8a587e66-60e3-47ac-9de0-44ed3092569f)

After : 
![image](https://media.github.cds.internal.unity3d.com/user/2768/files/05cbdd70-b7bd-4a2b-8263-082e0039b5ba)
![image](https://media.github.cds.internal.unity3d.com/user/2768/files/f223dd2a-fa51-4a57-9223-cd3c3a31fc33)